Example improvement
