+++ example/lib/exts.dart
@@ -1,0 +1,25 @@
+//
+//
+//
+
+import 'package:flutter/material.dart';
+
+extension LKExampleExt on BuildContext {
+  //
+  Future<void> showErrorDialog(dynamic exception) async {
+    //
+    await showDialog<void>(
+      context: this,
+      builder: (ctx) => AlertDialog(
+        title: const Text('Error'),
+        content: Text(exception.toString()),
+        actions: [
+          TextButton(
+            onPressed: () => Navigator.pop(ctx),
+            child: const Text('OK'),
+          )
+        ],
+      ),
+    );
+  }
+}
