+++ pubspec.yaml
@@ -13,7 +13,7 @@ dependencies:
   flutter:
     sdk: flutter
   flutter_webrtc:
-    path: ../flutter-webrtc/
+    path: ../../repos_livekit/flutter-webrtc/
     # git:
     #   url: https://github.com/flutter-webrtc/flutter-webrtc
     #   ref: master
